requirements.txt
# CRM Web App Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
SQLAlchemy==2.0.21
python-dotenv==1.0.0
email-validator==2.0.0
anthropic==0.18.1
openai==1.13.3
gunicorn==21.2.0